bug fix for end of file
Fixed following issues,
1. When end of file is called, which xbmc does to check the proper
length, axel could end the stream due to a bug.
2. if the chunk that asked for is in process, it could be in middle of
nowhere causing axel to stop working.
3. remove t0mmo common as its not needed and not part of Gotham anymore
4a. when seek or anything else happens, XBMC calls the new byte
location, so if the download is already going on then we should check
the filesize from downloading component as some of the sites mind
calling a Head multiple times.
4b. if 503 occurs, we conitnue downloading, probably with some
experience we would know the # of threads per server. BU stops 2
connection but we will see.
5. print cleanup, there still few but trying to clean up.